split partial handler into several files
