+++ README.md
@@ -1,0 +1,14 @@
+# TinyTypeMapper
+
+> **Mapping made simple**
+
+
+## Test with coverage
+
+```sh
+dotnet test /p:CollectCoverage=true /p:CoverletOutputFormat=lcov /p:CoverletOutput=./lcov
+```
+
+## Built with
+
+[.NET Core](https://docs.microsoft.com/en-us/dotnet/core/about)
